LibWeb/CSS: Extract ValueComparingRefPtr types into their own header
This is to prevent a circular dependency, but also, there's nothing
about this type that is CSS or LibWeb-specific, so let's put this in AK.
