+++ Libraries/LibWeb/CSS/StyleValues/StyleValue.h
@@ -16,6 +16,7 @@
 #include <AK/String.h>
 #include <AK/StringBuilder.h>
 #include <AK/StringView.h>
+#include <AK/ValueComparingRefPtr.h>
 #include <AK/Vector.h>
 #include <AK/WeakPtr.h>
 #include <LibGfx/Color.h>
@@ -98,58 +99,6 @@ namespace Web::CSS {
     __ENUMERATE_CSS_STYLE_VALUE_TYPE(URL, url, URLStyleValue)                                                         \
     __ENUMERATE_CSS_STYLE_VALUE_TYPE(ValueList, value_list, StyleValueList)                                           \
     __ENUMERATE_CSS_STYLE_VALUE_TYPE(ViewFunction, view_function, ViewFunctionStyleValue)
-
-template<typename T>
-struct ValueComparingNonnullRefPtr : public NonnullRefPtr<T> {
-    using NonnullRefPtr<T>::NonnullRefPtr;
-
-    ValueComparingNonnullRefPtr(NonnullRefPtr<T> const& other)
-        : NonnullRefPtr<T>(other)
-    {
-    }
-
-    ValueComparingNonnullRefPtr(NonnullRefPtr<T>&& other)
-        : NonnullRefPtr<T>(move(other))
-    {
-    }
-
-    bool operator==(ValueComparingNonnullRefPtr const& other) const
-    {
-        return this->ptr() == other.ptr() || this->ptr()->equals(*other);
-    }
-
-private:
-    using NonnullRefPtr<T>::operator==;
-};
-
-template<typename T>
-struct ValueComparingRefPtr : public RefPtr<T> {
-    using RefPtr<T>::RefPtr;
-
-    ValueComparingRefPtr(RefPtr<T> const& other)
-        : RefPtr<T>(other)
-    {
-    }
-
-    ValueComparingRefPtr(RefPtr<T>&& other)
-        : RefPtr<T>(move(other))
-    {
-    }
-
-    template<typename U>
-    bool operator==(ValueComparingNonnullRefPtr<U> const& other) const
-    {
-        return this->ptr() == other.ptr() || (this->ptr() && this->ptr()->equals(*other));
-    }
-
-    bool operator==(ValueComparingRefPtr const& other) const
-    {
-        return this->ptr() == other.ptr() || (this->ptr() && other.ptr() && this->ptr()->equals(*other));
-    }
-
-private:
-    using RefPtr<T>::operator==;
-};
 
 struct ColorResolutionContext {
     Optional<PreferredColorScheme> color_scheme;
